feat(message): move emojis to emoji.go
